+++ src/balance/balance.service.ts
@@ -1,4 +1,4 @@
-import { Inject, Injectable, NotFoundException } from '@nestjs/common';
+import { Inject, Injectable, Logger, NotFoundException } from '@nestjs/common';
 import { fromPairs } from 'lodash';
 
 import { BalanceFetcherRegistry } from './balance-fetcher.registry';
@@ -6,6 +6,8 @@ import { GetBalancesQuery } from './dto/get-balances-query.dto';
 
 @Injectable()
 export class BalanceService {
+  private logger = new Logger(BalanceService.name);
+
   constructor(@Inject(BalanceFetcherRegistry) private readonly balanceFetcherRegistry: BalanceFetcherRegistry) {}
 
   async getBalances({ appId, addresses, network }: GetBalancesQuery & { appId: string }) {
@@ -16,7 +18,10 @@ export class BalanceService {
           fetcher
             .getBalances(address)
             .then(balance => [address, balance])
-            .catch(e => [address, { error: e }]),
+            .catch(e => {
+              this.logger.error(`Failed to fetch balance for ${appId} on network ${network}: ${e.stack}`);
+              return [address, { error: e.message }];
+            }),
         ),
       );